innotter/pyproject.toml
[tool.poetry]
name = "innotter"
version = "0.1.0"
description = "twitter clone"
authors = ["Sheshkon <sheskoalexei@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.5.2"
certifi = "2022.5.18.1"
charset-normalizer = "2.0.12"
coreapi = "2.3.3"
coreschema = "0.0.4"
Django = "4.0.5"
django-rest-swagger = "2.2.0"
djangorestframework = "3.13.1"
idna = "3.3"
itypes = "1.2.0"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.1"
openapi-codec = "1.3.2"
PyJWT = "2.4.0"
python-dotenv = "0.20.0"
pytz = "2022.1"
requests = "2.28.0"
simplejson = "3.17.6"
sqlparse = "0.4.2"
uritemplate = "4.1.1"
urllib3 = "1.26.9"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
coverage = "^6.4.1"
pytest = "^7.1.2"
pytest-django = "^4.5.2"
celery = "^5.2.7"
pytest-mock = "^3.8.2"
boto3 = "^1.24.29"
awscli-local = "^0.20"
awscli = "^1.25.31"
Pillow = "^9.2.0"
pika = "^1.3.0"
django-celery-beat = "^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
